Flask Web App/requirements.txt

tensorflow==2.3.1
Pillow
Flask
requests
numpy


